frontend/requirements.txt
streamlit>=1.24.0
opencv-python>=4.8.0
pillow>=9.5.0
pandas>=2.0.3 